requirements.txt
quart
line-bot-sdk
playwright
beautifulsoup4
gunicorn
asyncio